backend/requirements.txt
dvc==3.12.0
opencv-python==4.8.0.74
opencv-contrib-python
parameterized==0.9.0
Pillow==9.5.0
ultralytics==8.0.147
flask
flask_pymongo
flask_cors